Moved APIHandler.ts in advance of change to client/provider interaction. Changing as two commits so that file can be moved whilst preserving history.
